feat(chat): implement session-scoped error handling and refactor chat composer
- Add `visibleChatError` function to manage session-specific error messages.
- Update `useChat` hook to track errors by session and global errors.
- Refactor chat composer to separate concerns and improve readability.
- Introduce `ChatComposer` component to encapsulate chat input logic.
- Adjust `ComposerTrailingControls` for improved button sizes and styles.
- Enhance tests for error visibility based on active session.
